feat: 결제 금액 타입 변경 - #34

+++ src/main/java/com/permitseoul/permitserver/domain/payment/api/dto/PaymentResponse.java
@@ -1,10 +1,12 @@
 package com.permitseoul.permitserver.domain.payment.api.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentResponse(
         String paymentKey,
         String orderId,
         String currency,
-        int totalAmount,
+        BigDecimal totalAmount,
         String requestedAt,
         String approvedAt
 ) {
