+++ src/main/java/com/permitseoul/permitserver/domain/reservation/api/dto/PaymentConfirmRequest.java
@@ -3,6 +3,8 @@ package com.permitseoul.permitserver.domain.reservation.api.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record PaymentConfirmRequest(
 
         @NotBlank(message = "orderId는  없으면 안됩니다.")
@@ -12,5 +14,5 @@ public record PaymentConfirmRequest(
         String paymentKey,
 
         @Positive(message = "totalAmount는 0보다 커야합니다.")
-        int totalAmount
+        BigDecimal totalAmount
 ) { }
